% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDensityIncidence.R
\name{buscar_parelles_dtindex}
\alias{buscar_parelles_dtindex}
\title{Buscar parelles dtindex}
\usage{
buscar_parelles_dtindex(dt, eventcontrol, reemplacement, Ncontrols)
}
\arguments{
\item{dt}{dt}

\item{eventcontrol}{eventcontrol}

\item{reemplacement}{reemplacement}

\item{Ncontrols}{Ncontrols}
}
\value{
base de dades
}
\description{
Buscar parelles dtindex
}
