% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{calcular_proporcio}
\alias{calcular_proporcio}
\title{calcular_proporcio}
\usage{
calcular_proporcio(dt = dades, factor = "canvi612M.glicadaCAT2")
}
\arguments{
\item{dt}{dt}

\item{factor}{factor}
}
\description{
calcular_proporcio
}
