% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductor1_test.R
\docType{data}
\name{conductor1_test}
\alias{conductor1_test}
\title{Conductor de la base de dades Plana}
\format{
Una data frame amb 15 files i 9 variables
\describe{
\item{camp}{camp}
\item{dtindex}{dtindex}
\item{descripcio}{descripcio}
\item{descripcio2}{descripcio2}
\item{factor}{factor}
\item{dates}{dates}
\item{recode}{recode}
\item{rang_valid}{rang_valid}
\item{taula1}{taula1}
\item{taula2}{taula2}
\item{ajust4}{ajust4}
}
}
\usage{
conductor1_test
}
\description{
Conductor de la base de dades Plana
}
\keyword{datasets}
