% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{covariate_plot}
\alias{covariate_plot}
\title{Covariate plot}
\usage{
covariate_plot(
  dt = "m.out",
  vars_remove = NULL,
  etiquetar = F,
  subtitle = "oGLD vs SGLT-2i group",
  ...
)
}
\arguments{
\item{dt}{dt}

\item{vars_remove}{vars_remove}

\item{etiquetar}{etiquetar}

\item{subtitle}{subtitle}

\item{...}{Altres parametres}
}
\description{
Covariate plot
}
