% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{covariate_plot_dades}
\alias{covariate_plot_dades}
\title{covariate_plot_dades}
\usage{
covariate_plot_dades(
  dt = "dt_total",
  var = "name",
  stat = "stat",
  title = "Covariate plot \\n oGLD vs SGLT-2i group",
  labx = "Standardized mean difference"
)
}
\arguments{
\item{dt}{dt}

\item{var}{var}

\item{stat}{stat}

\item{title}{title}

\item{labx}{labx}
}
\description{
covariate_plot_dades
}
