% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_test.R
\docType{data}
\name{dat_test}
\alias{dat_test}
\title{Base de dades per fer Matching}
\format{
Una data frame amb 55 files i 8 variables
\describe{
\item{event}{cas}
\item{idp}{idp}
\item{sex}{sexe}
\item{byear}{any}
\item{diabetes}{diabetis}
\item{heartdis}{cardiopatia}
\item{dtindex_case}{data index del cas}
\item{dtindex_control}{data censura}
}
}
\usage{
dat_test
}
\description{
Base de dades per fer Matching
}
\keyword{datasets}
