% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{dt_index_data_semirandom}
\alias{dt_index_data_semirandom}
\title{Dt_index_data_semirandom}
\usage{
dt_index_data_semirandom(
  dt = "PACIENTS",
  dt.variables = "VARIABLES",
  codi = "EK201"
)
}
\arguments{
\item{dt}{dt}

\item{dt.variables}{dt.variables}

\item{codi}{codi}
}
\description{
Dt_index_data_semirandom
}
