% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{etiquetar_model}
\alias{etiquetar_model}
\title{Etiquetar_model}
\usage{
etiquetar_model(
  model = "kkk",
  taulavariables = "variables_R.xls",
  camp = "camp",
  camp_descripcio = "descripcio",
  ...
)
}
\arguments{
\item{model}{Model}

\item{taulavariables}{Conductor}

\item{camp}{Camp}

\item{camp_descripcio}{Camp DESCRIPCIO}

\item{...}{Altres parametres}
}
\value{
Etiquetar model
}
\description{
Etiquetar_model
}
