% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altresFuncions.R
\name{etiquetar_taula}
\alias{etiquetar_taula}
\title{Etiquetar una Taula}
\usage{
etiquetar_taula(
  taula = "resumtotal",
  camp = "variable",
  taulavariables = "variables_R.xls",
  camp_descripcio = "descripcio",
  idcamp = "camp"
)
}
\arguments{
\item{taula}{Taula}

\item{camp}{Camp}

\item{taulavariables}{xxx}

\item{camp_descripcio}{xxx}

\item{idcamp}{xxx}
}
\value{
Nova Taula
}
\description{
Etiquetar una Taula
}
