% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure.dif.proporcions}
\alias{extreure.dif.proporcions}
\title{Extreure dif proporcions}
\usage{
extreure.dif.proporcions(
  dades,
  outcome = "Prediabetes",
  ref_cat = NA,
  grups = "Sex"
)
}
\arguments{
\item{dades}{dades}

\item{outcome}{outcome}

\item{ref_cat}{ref_cat}

\item{grups}{grups}
}
\description{
Extreure dif proporcions
}
