% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altresFuncions.R
\name{extreure.variables}
\alias{extreure.variables}
\title{Extreure variables seleccionades}
\usage{
extreure.variables(
  taula = "table1",
  taulavariables = "variables_R.xls",
  variable_camp = "camp",
  dt = NA,
  ...
)
}
\arguments{
\item{taula}{Camp a on hi han les variables}

\item{taulavariables}{Conductor}

\item{variable_camp}{Variable camp del Conductor}

\item{dt}{Base de dades}

\item{...}{altres funcions}
}
\value{
un vectora amb les variables seleccionades
}
\description{
Extreure variables seleccionades a partir d un conductor
}
\examples{

k1<-extreure.variables(taula="taula1",taulavariables=conductor1,variable_camp="camp",dt=NA)
k2<-extreure.variables(taula="taula1",taulavariables=conductor1,variable_camp="camp",dt=dt_plana)
k1
k2

}
