% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_coef_cmprisk}
\alias{extreure_coef_cmprisk}
\title{Extreure coef cmprisk}
\usage{
extreure_coef_cmprisk(model_cmrisk)
}
\arguments{
\item{model_cmrisk}{model_cmrisk}
}
\description{
Extreure coef cmprisk
}
