% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_coef_glm}
\alias{extreure_coef_glm}
\title{Extreure coef glm}
\usage{
extreure_coef_glm(
  dt = "dades",
  outcomes = "OFT_WORST",
  x = "DM",
  z = "",
  taulavariables = "variables_R.xls"
)
}
\arguments{
\item{dt}{dt}

\item{outcomes}{outcomes}

\item{x}{x}

\item{z}{z}

\item{taulavariables}{taulavariables}
}
\description{
Extreure coef glm
}
