% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_coef_glm_mi}
\alias{extreure_coef_glm_mi}
\title{Extreure coef glm mi}
\usage{
extreure_coef_glm_mi(
  dt = "tempData",
  outcome = "valor612M.GLICADA",
  x = "SEXE",
  v.ajust = "",
  level_conf = 0.95
)
}
\arguments{
\item{dt}{dt}

\item{outcome}{outcome}

\item{x}{x}

\item{v.ajust}{v.ajust}

\item{level_conf}{level_conf}
}
\description{
Extreure coef glm mi
}
