% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_coef_glm_v2}
\alias{extreure_coef_glm_v2}
\title{extreure coef glm v2}
\usage{
extreure_coef_glm_v2(
  dt = "dades",
  outcome = "OFT_WORST",
  x = "DM",
  v.ajust = "",
  level_conf = 0.95
)
}
\arguments{
\item{dt}{dt}

\item{outcome}{outcome}

\item{x}{x}

\item{v.ajust}{v.ajust}

\item{level_conf}{level_conf}
}
\description{
extreure coef glm v2
}
