% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_coef_mice_estrats}
\alias{extreure_coef_mice_estrats}
\title{Extreure coef mice estrats}
\usage{
extreure_coef_mice_estrats(
  tempData,
  data_list,
  X = c("bmi", "hyp"),
  Y = "chl",
  grups = "age",
  logit = F
)
}
\arguments{
\item{tempData}{tempData}

\item{data_list}{data_list}

\item{X}{X}

\item{Y}{Y}

\item{grups}{grups}

\item{logit}{logit}
}
\description{
Extreure coef mice estrats
}
