% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_cor}
\alias{extreure_cor}
\title{Extreure cor}
\usage{
extreure_cor(
  var1 = "CD36",
  var = "quantis",
  d = "dades",
  taulavariables = "VARIABLES.xls",
  ...
)
}
\arguments{
\item{var1}{var1}

\item{var}{var}

\item{d}{d}

\item{taulavariables}{taulavariables}

\item{...}{Altres parametres}
}
\description{
Extreure cor
}
