% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_cor_multi}
\alias{extreure_cor_multi}
\title{Extreure cor multi}
\usage{
extreure_cor_multi(
  dades = dt,
  llistavar1 = c("Age", "BMI"),
  llistavar2 = c("Large_PER_HDL", "Medium_HDL_P_molL"),
  etiquetar = F,
  coductor_variables = "conductor_variables",
  method = "circle",
  ...
)
}
\arguments{
\item{dades}{dades}

\item{llistavar1}{llistavar1}

\item{llistavar2}{llistavar2}

\item{etiquetar}{etiquetar}

\item{coductor_variables}{coductor_variables}

\item{method}{method}

\item{...}{Altres parametres}
}
\description{
Extreure cor multi
}
