% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_cuminc_cmprisk}
\alias{extreure_cuminc_cmprisk}
\title{extreure_cuminc_cmprisk}
\usage{
extreure_cuminc_cmprisk(
  dt = dades,
  event = "amputacio_cat",
  temps = "t_lliure_amputa",
  competitiu = "Ha_muerto",
  codievent = "Yes",
  group = NULL,
  strata = NULL
)
}
\arguments{
\item{dt}{dt}

\item{event}{event}

\item{temps}{temps}

\item{competitiu}{competitiu}

\item{codievent}{codievent}

\item{group}{group}

\item{strata}{strata}
}
\description{
extreure_cuminc_cmprisk
}
