% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_HRFG}
\alias{extreure_HRFG}
\title{extreure_HRFG}
\usage{
extreure_HRFG(
  event = "exitusCV",
  temps = "temps_seguiment",
  grup = "diabetis",
  eventcompetitiu = "exitus",
  dt = "dades",
  covariables = NA,
  codievent = "Si",
  refcat = NA,
  cluster = ""
)
}
\arguments{
\item{event}{event}

\item{temps}{temps}

\item{grup}{grup}

\item{eventcompetitiu}{eventcompetitiu}

\item{dt}{dt}

\item{covariables}{covariables}

\item{codievent}{codievent}

\item{refcat}{refcat}

\item{cluster}{cluster}
}
\description{
extreure_HRFG
}
