% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_model_cmprisk}
\alias{extreure_model_cmprisk}
\title{Extreure model cmprisk}
\usage{
extreure_model_cmprisk(
  dt = dades,
  event = "amputacio_cat",
  temps = "t_lliure_amputa",
  competitiu = "Ha_muerto",
  codievent = "Yes",
  covariables = c("Sexo", "Edad", "Tabaquismo")
)
}
\arguments{
\item{dt}{dt}

\item{event}{event}

\item{temps}{temps}

\item{competitiu}{competitiu}

\item{codievent}{codievent}

\item{covariables}{covariables}
}
\description{
Extreure model cmprisk
}
