% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_model_logistic}
\alias{extreure_model_logistic}
\title{Extreure model logistic}
\usage{
extreure_model_logistic(
  x = "OS4_GSK",
  y = "canvi6M.glipesCAT2",
  taulavariables = "conductorvariables",
  dades = dades,
  elimina = c("IDP"),
  a = "",
  valor_outcome = "Yes",
  conditional = F,
  strata = "caseid"
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{taulavariables}{taulavariables}

\item{dades}{dades}

\item{elimina}{elimina}

\item{a}{a}

\item{valor_outcome}{valor_outcome}

\item{conditional}{conditional}

\item{strata}{strata}
}
\description{
Extreure model logistic
}
