% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_Pglobal_SigTest}
\alias{extreure_Pglobal_SigTest}
\title{Extreure Pglobal SigTest}
\usage{
extreure_Pglobal_SigTest(
  dt = dades,
  vars_pre = vars_pre,
  vars_post = vars_post
)
}
\arguments{
\item{dt}{dt}

\item{vars_pre}{vars_pre}

\item{vars_post}{vars_post}
}
\description{
Extreure Pglobal SigTest
}
