% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_resum_outcomes_imputation}
\alias{extreure_resum_outcomes_imputation}
\title{Extreure resum outcomes imputation}
\usage{
extreure_resum_outcomes_imputation(
  dades_long = dades,
  outcome = "HBA1C.dif324m",
  grups = "grup",
  v.ajust = c("sexe", "edat"),
  level_conf = 0.95
)
}
\arguments{
\item{dades_long}{dades_long}

\item{outcome}{outcome}

\item{grups}{grups}

\item{v.ajust}{v.ajust}

\item{level_conf}{level_conf}
}
\description{
Extreure resum outcomes imputation
}
