% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{fores.plot.v4}
\alias{fores.plot.v4}
\title{Forest.plot.v4}
\usage{
fores.plot.v4(
  dadesmodel = "dt_outHR",
  label = "etiqueta",
  mean = "estimate",
  lower = "Linf",
  upper = "Lsup",
  label_X = "OR (95\% CI)",
  intercept = 1,
  nivell = "outcome",
  factor = "Event",
  label_Xvertical = "Cardiovascular event",
  title = "Forest plot of hazard hatios and confidence interval (95\%CI)",
  subtitle = "Sensitivity analysis",
  caption = "",
  label_Favors = "Favors SGLT-2        Favors oGLD-2"
)
}
\arguments{
\item{dadesmodel}{dadesmodel}

\item{label}{label}

\item{mean}{mean}

\item{lower}{lower}

\item{upper}{upper}

\item{label_X}{label_X}

\item{intercept}{intercept}

\item{nivell}{nivell}

\item{factor}{factor}

\item{label_Xvertical}{label_Xvertical}

\item{title}{title}

\item{subtitle}{subtitle}

\item{caption}{caption}

\item{label_Favors}{label_Favors}
}
\description{
Forest.plot.v4
}
