% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{forest.plot.HR}
\alias{forest.plot.HR}
\title{forest.plot.HR}
\usage{
forest.plot.HR(
  dadesmodel,
  label = "Categoria",
  mean = "estimate",
  lower = "Linf",
  upper = "Lsup",
  label_X = "OR (95\% CI)",
  intercept = 1,
  nivell = "outcome",
  factor1 = "type",
  color = F,
  label_Xvertical = "Cardiovascular event",
  nolabels = TRUE,
  title = "Forest plot of hazard hatios and confidence interval (95\%CI)",
  label_Favors = "Favors SGLT-2        Favors oGLD-2"
)
}
\arguments{
\item{dadesmodel}{dadesmodel}

\item{label}{label}

\item{mean}{mean}

\item{lower}{lower}

\item{upper}{upper}

\item{label_X}{label_X}

\item{intercept}{intercept}

\item{nivell}{nivell}

\item{factor1}{factor1}

\item{color}{color}

\item{label_Xvertical}{label_Xvertical}

\item{nolabels}{nolabels}

\item{title}{title}

\item{label_Favors}{label_Favors}
}
\description{
forest.plot.HR
}
