% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{forest.plot.v2}
\alias{forest.plot.v2}
\title{forest.plot.v2}
\usage{
forest.plot.v2(
  dadesmodel = "ramo",
  label = "Categoria",
  mean = "OR",
  lower = "Linf",
  upper = "Lsup",
  label_X = "OR (95\% CI)",
  intercept = 1
)
}
\arguments{
\item{dadesmodel}{dadesmodel}

\item{label}{label}

\item{mean}{mean}

\item{lower}{lower}

\item{upper}{upper}

\item{label_X}{label_X}

\item{intercept}{intercept}
}
\description{
forest.plot.v2
}
