% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{forest.plot.v3}
\alias{forest.plot.v3}
\title{Forest.plot.v3}
\usage{
forest.plot.v3(
  dadesmodel = "dt_estimacions",
  label = "Categoria",
  mean = "estimate",
  lower = "Linf",
  upper = "Lsup",
  label_X = "OR (95\% CI)",
  intercept = 0,
  nivell = "outcome",
  factor1 = "type",
  factor2 = "datos",
  color = TRUE
)
}
\arguments{
\item{dadesmodel}{dadesmodel}

\item{label}{label}

\item{mean}{mean}

\item{lower}{lower}

\item{upper}{upper}

\item{label_X}{label_X}

\item{intercept}{intercept}

\item{nivell}{nivell}

\item{factor1}{factor1}

\item{factor2}{factor2}

\item{color}{color}
}
\description{
Forest.plot.v3
}
