% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{formula.LOGIT}
\alias{formula.LOGIT}
\title{Formula LOGIT}
\usage{
formula.LOGIT(
  x = "taula1",
  y = "resposta",
  eliminar = c("IDP"),
  a = "",
  taulavariables = "variables.xls"
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{eliminar}{eliminar}

\item{a}{a}

\item{taulavariables}{taulavariables}
}
\description{
Formula LOGIT
}
