% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{formula_compare}
\alias{formula_compare}
\title{FORMULA A PARTIR DE VARIABLES}
\usage{
formula_compare(
  x = "taula1",
  y = "grup",
  elimina = c("IDP"),
  taulavariables = "variables_R.xls",
  dt = "No",
  ...
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{elimina}{elimina}

\item{taulavariables}{taulavariables}

\item{dt}{dt}

\item{...}{Altres parametres}
}
\value{
FORMULA
}
\description{
FORMULA A PARTIR DE VARIABLES:compte l'Ordre que està posada en el conductor taulavariables
}
