% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altresFuncions.R
\name{formula_text}
\alias{formula_text}
\title{Formula Text.}
\usage{
formula_text(
  x = "taula1",
  y = "resposta",
  eliminar = c("IDP"),
  a = "",
  taulavariables,
  dt = NA
)
}
\arguments{
\item{x}{Indicador de cadena del camp del fitxer extern amb indicador de la posició de la fórmula (1,2....)}

\item{y}{Indicador de resposta de cadena (default "."): ". ~ x1 +x2 + x3 "}

\item{eliminar}{Caracter que indica si alguna variable ha de ser eliminada}

\item{a}{Caracter , indica si cal afegir alguna variable en la primera posició}

\item{taulavariables}{Conductor , a on triem l'orde de les variables de la formula}

\item{dt}{data.frame o base de dades}
}
\value{
formula
}
\description{
Retorna Data.frame or tibble amb
}
\examples{

Hmisc::label(iris)
conductor_iris<-data.frame(camp=names(iris),formu=c(1,2,3,4,5))
formula_iris1<-formula_text(x="formu",y="RESPOSTA",taulavariables = conductor_iris)
formula_iris1
}
