% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altresFuncions.R
\name{formula_vector}
\alias{formula_vector}
\title{Formula vector}
\usage{
formula_vector(
  vector = c("sex", "age", "age"),
  y = "y",
  logit = F,
  eliminar = NA
)
}
\arguments{
\item{vector}{Vector amb les variables}

\item{y}{Variable dependents}

\item{logit}{Logit}

\item{eliminar}{Aquelles variables per eliminar}
}
\value{
Formula_vector
}
\description{
Formula vector, vector amb les variables i y
}
\examples{
vector=c("sex","age","age")
formula_iris3<-formula_vector(vector,y="y")
formula_iris3
}
