% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{formulaCOX}
\alias{formulaCOX}
\title{formula COX}
\usage{
formulaCOX(
  x = "",
  event = "event",
  temps = "temps",
  elimina = "",
  cluster = "",
  a = "",
  taulavariables = "variables.xls",
  codievent = "1"
)
}
\arguments{
\item{x}{x}

\item{event}{event}

\item{temps}{temps}

\item{elimina}{elimina}

\item{cluster}{cluster}

\item{a}{a}

\item{taulavariables}{taulavariables}

\item{codievent}{codievent}
}
\description{
formula COX
}
