% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{Gaps}
\alias{Gaps}
\title{Gaps}
\usage{
Gaps(dt = dades, K = 14, Nmostra = 10, finestraX = c(NA, NA), llavor = 123)
}
\arguments{
\item{dt}{dt}

\item{K}{K}

\item{Nmostra}{Nmostra}

\item{finestraX}{finestraX}

\item{llavor}{llavor}
}
\description{
Gaps
}
