% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{generar_Surv}
\alias{generar_Surv}
\title{generar Surv}
\usage{
generar_Surv(dt, event, dtindex = "dtindex", dtsortida = "sortida")
}
\arguments{
\item{dt}{dt}

\item{event}{event}

\item{dtindex}{dtindex}

\item{dtsortida}{dtsortida}
}
\description{
generar_Surv
}
