% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{generar_Surv_to_column}
\alias{generar_Surv_to_column}
\title{generar Surv to column}
\usage{
generar_Surv_to_column(
  dt = "dadestotal",
  event = "EV.AVC_DIC",
  temps = "EV.AVC_temps",
  codievent = 1
)
}
\arguments{
\item{dt}{dt}

\item{event}{event}

\item{temps}{temps}

\item{codievent}{codievent}
}
\description{
generar Surv to column
}
