% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{generar_taula_variables_formula}
\alias{generar_taula_variables_formula}
\title{Generar taula variables formula}
\usage{
generar_taula_variables_formula(formu = "AnyPlaqueBasal~CD5L", dades = dt)
}
\arguments{
\item{formu}{formu}

\item{dades}{dades}
}
\description{
Generar taula variables formula
}
