% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{ggplot_grups}
\alias{ggplot_grups}
\title{GGPLOT grups}
\usage{
ggplot_grups(Y = "DIS_estatina", dt = dades, X = "edat", grup = "sexe")
}
\arguments{
\item{Y}{Y}

\item{dt}{dt}

\item{X}{X}

\item{grup}{grup}
}
\description{
GGPLOT grups
}
