% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{HR.COX.CRU}
\alias{HR.COX.CRU}
\title{HR.COX.CRU}
\usage{
HR.COX.CRU(
  x = "lipos",
  event = "EVENT_MCV",
  t = "temps_exitus",
  e = "",
  d = "dadesDF",
  variables = "variables_R.xls",
  evento = "Si"
)
}
\arguments{
\item{x}{x}

\item{event}{event}

\item{t}{t}

\item{e}{e}

\item{d}{d}

\item{variables}{variables}

\item{evento}{evento}
}
\description{
HR.COX.CRU
}
