% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{HR.COX}
\alias{HR.COX}
\title{HR.COX}
\usage{
HR.COX(
  x = "",
  event = "EV.INSUF_CARD",
  t = "tmp_insuf_card",
  e = "",
  d = "dadesDF",
  taulavariables = "variables.xls",
  c = "",
  ...
)
}
\arguments{
\item{x}{x}

\item{event}{event}

\item{t}{t}

\item{e}{e}

\item{d}{d}

\item{taulavariables}{taulavariables}

\item{c}{c}

\item{...}{Altres parametres}
}
\description{
HR.COX
}
