% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{HRadj}
\alias{HRadj}
\title{HRadj}
\usage{
HRadj(
  x = "v.ajust",
  event = "EV.INSUF_CARD",
  t = "tmp_insuf_card",
  e = "",
  c = "",
  d = "dadesDF",
  taulavariables = "variables.xls",
  codievent = "1"
)
}
\arguments{
\item{x}{x}

\item{event}{event}

\item{t}{t}

\item{e}{e}

\item{c}{c}

\item{d}{d}

\item{taulavariables}{taulavariables}

\item{codievent}{codievent}
}
\description{
HRadj
}
