% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{HRestratificats}
\alias{HRestratificats}
\title{HRestratificats}
\usage{
HRestratificats(
  event = "exitus",
  t = "temps",
  tipo = "v.ajust",
  c = "",
  taulavariables = "variables.xls"
)
}
\arguments{
\item{event}{event}

\item{t}{t}

\item{tipo}{tipo}

\item{c}{c}

\item{taulavariables}{taulavariables}
}
\description{
HRestratificats
}
