% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{llista.compare.Ys}
\alias{llista.compare.Ys}
\title{Llistat de taules a partir de Llista de factors de Y's}
\usage{
llista.compare.Ys(
  dt = "dades",
  llista.y = c("CODGLP1", "CKDEPI_cat2"),
  llista.x = c("canvi612.pes.perc", "canvi612M.pes"),
  show.ratio = F,
  byrow = T,
  show.n = T,
  show.all = T,
  show.descr = T,
  digits = NA,
  digits.ratio = NA,
  hide.no = c("NA", "No"),
  ref.no = NA
)
}
\arguments{
\item{dt}{dt}

\item{llista.y}{llista.y}

\item{llista.x}{llista.x}

\item{show.ratio}{show.ratio}

\item{byrow}{byrow}

\item{show.n}{show.n}

\item{show.all}{show.all}

\item{show.descr}{show.descr}

\item{digits}{digits}

\item{digits.ratio}{digits.ratio}

\item{hide.no}{hide.no}

\item{ref.no}{ref.no}
}
\value{
llista
}
\description{
Llistat de taules a partir de Llista de factors de Y's i em retorna una llista de taules
}
