% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{llistadetaules.compare}
\alias{llistadetaules.compare}
\title{llistadetaules.compare}
\usage{
llistadetaules.compare(
  tablero = c("taula1", "taula2", "taula3", "taula4", "taula5"),
  y = "sexe",
  variables = "variables.xls",
  dt = dades
)
}
\arguments{
\item{tablero}{tablero}

\item{y}{y}

\item{variables}{variables}

\item{dt}{dt}
}
\description{
llistadetaules.compare
}
