% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Macedonia.R
\docType{package}
\name{Macedonia}
\alias{Macedonia}
\title{Macedonia Package}
\description{
Macedonia és un Package que conté les funcions necessàries per treballar en bases de dades mèdiques
i que formen part dels Packages Platan i FormatGe.
}
\author{
Jordi Real,Ray Puig \email{-jordireal@gmail.com-}
}
