% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altresFuncions.R
\name{make_dummies}
\alias{make_dummies}
\title{Genera dummis}
\usage{
make_dummies(dt, variable, prefix = " ")
}
\arguments{
\item{dt}{Base de dades}

\item{variable}{Variables}

\item{prefix}{Prefix}
}
\value{
Retorna dummis
}
\description{
Recodificar. Genera dummis (0/1) a partir d'una variable del data frame
}
\examples{

sexe_dummis<-make_dummies(dt_plana,"sexe",prefix="pref.")
sexe_dummis

}
