% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{MAP_ggplot}
\alias{MAP_ggplot}
\title{MAP ggplot}
\usage{
MAP_ggplot(
  dades = dt,
  datainicial = "data",
  datafinal = "datafi",
  id = "idp_temp",
  grup_color = NA,
  grup_linea = NA,
  lim_inf = -Inf,
  lim_sup = Inf,
  add_point = NA
)
}
\arguments{
\item{dades}{dades}

\item{datainicial}{datainicial}

\item{datafinal}{datafinal}

\item{id}{id}

\item{grup_color}{grup_color}

\item{grup_linea}{grup_linea}

\item{lim_inf}{lim_inf}

\item{lim_sup}{lim_sup}

\item{add_point}{add_point}
}
\description{
MAP ggplot
}
