% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{MAP_ggplot_univariant}
\alias{MAP_ggplot_univariant}
\title{MAP_ggplot_univariant}
\usage{
MAP_ggplot_univariant(
  dades = dt,
  datainicial = "data",
  datafinal = "datafi",
  id = "idp_temp",
  Nmostra = 10,
  add_point = NA,
  add_final = NA,
  set_seed = 123
)
}
\arguments{
\item{dades}{dades}

\item{datainicial}{datainicial}

\item{datafinal}{datafinal}

\item{id}{id}

\item{Nmostra}{Nmostra}

\item{add_point}{add_point}

\item{add_final}{add_final}

\item{set_seed}{set_seed}
}
\description{
MAP_ggplot_univariant
}
