% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{MAP_punts_ggplot}
\alias{MAP_punts_ggplot}
\title{MAP punts ggplot}
\usage{
MAP_punts_ggplot(
  dt = "mostra50",
  id = "idp",
  datainicial = "dat",
  val = "val",
  grup_color = "agr",
  Nmostra = Inf,
  llavor = 123,
  finestraX = c(-Inf, +Inf),
  id_AGG = F
)
}
\arguments{
\item{dt}{dt}

\item{id}{id}

\item{datainicial}{datainicial}

\item{val}{val}

\item{grup_color}{grup_color}

\item{Nmostra}{Nmostra}

\item{llavor}{llavor}

\item{finestraX}{finestraX}

\item{id_AGG}{id_AGG}
}
\description{
MAP punts ggplot
}
