% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{MAP_valor_ggplot}
\alias{MAP_valor_ggplot}
\title{MAP valor ggplot}
\usage{
MAP_valor_ggplot(
  dt = "mostra50",
  id = "idp",
  datainicial = "dat",
  val = "val",
  grup_color = "agr",
  Nmostra = 1,
  finestraX = c(-Inf, Inf),
  llavor = 123,
  title = "Evolucio de valors"
)
}
\arguments{
\item{dt}{dt}

\item{id}{id}

\item{datainicial}{datainicial}

\item{val}{val}

\item{grup_color}{grup_color}

\item{Nmostra}{Nmostra}

\item{finestraX}{finestraX}

\item{llavor}{llavor}

\item{title}{title}
}
\description{
MAP valor ggplot
}
