% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDensityIncidence.R
\name{match_density_incidence}
\alias{match_density_incidence}
\title{match density incidence}
\usage{
match_density_incidence(
  dt = "dt_poblacio",
  id = "idp",
  llistaPS = c("sexe"),
  eventcontrol = F,
  reemplacement = F,
  numcores = NA,
  Ncontrols = 1,
  seed = 123
)
}
\arguments{
\item{dt}{dt}

\item{id}{id}

\item{llistaPS}{llistaPS}

\item{eventcontrol}{eventcontrol}

\item{reemplacement}{reemplacement}

\item{numcores}{numcores}

\item{Ncontrols}{Ncontrols}

\item{seed}{seed}
}
\value{
match
}
\description{
match density incidence
}
\examples{

case <- c(rep(0,40),rep(1,15))
ptid <- paste0("P",1:55)
sex <- c(rep("fem",20),rep("mal",20),rep("fem",8),rep("mal",7))
byear <- c(rep(c(2020,2030),20),rep(2020,7),rep(2030,8))
case.Index <- c(seq(1,40,1),seq(5,47,3))
control.Index <- case.Index
diabetes <- seq(2,110,2)
heartdis <- seq(110,2,-2)
diabetes <- c(rep(1,55))
heartdis <- c(rep(100,55))

dat <- data.table::data.table(case,ptid,sex,byear,diabetes,heartdis,case.Index,control.Index)
}
