% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{matching_4grups}
\alias{matching_4grups}
\title{matching_4grups}
\usage{
matching_4grups(
  dt = dadesini,
  grups = "grup",
  vars_match = "matching",
  conductor = "vars_ilerbus.xls",
  caliper = 0.01
)
}
\arguments{
\item{dt}{dt}

\item{grups}{grups}

\item{vars_match}{vars_match}

\item{conductor}{conductor}

\item{caliper}{caliper}
}
\description{
matching_4grups
}
