% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{matching_case_control}
\alias{matching_case_control}
\title{matching_case_control}
\usage{
matching_case_control(
  dt = "PACIENTS",
  variables.ps = "llistaPS",
  dt_pacients_dindex = "BD_PAC_DINDEX"
)
}
\arguments{
\item{dt}{dt}

\item{variables.ps}{variables.ps}

\item{dt_pacients_dindex}{dt_pacients_dindex}
}
\description{
matching_case_control
}
