% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{mostreig_ids}
\alias{mostreig_ids}
\title{Mostreig ids}
\usage{
mostreig_ids(dt, id = "idp", n_mostra = 100, set_seed = 123)
}
\arguments{
\item{dt}{dt}

\item{id}{id}

\item{n_mostra}{n_mostra}

\item{set_seed}{set_seed}
}
\description{
Mostreig ids
}
