% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{OR.ajustats}
\alias{OR.ajustats}
\title{OR ajustats}
\usage{
OR.ajustats(
  x = "lipos",
  ajust = "V.ajust",
  y = "prediabetis",
  d = "dadestotal",
  taulavariables = "variables.xls"
)
}
\arguments{
\item{x}{x}

\item{ajust}{ajust}

\item{y}{y}

\item{d}{d}

\item{taulavariables}{taulavariables}
}
\description{
OR ajustats
}
