% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{plot.dispersio.reduccio}
\alias{plot.dispersio.reduccio}
\title{plot.dispersio.reduccio}
\usage{
\method{plot}{dispersio.reduccio}(dt = "dades", v.basal = "HBpreADD", v.final = "HBpostADD")
}
\arguments{
\item{dt}{dt}

\item{v.basal}{v.basal}

\item{v.final}{v.final}
}
\description{
plot.dispersio.reduccio
}
