% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{proporcio_grups_estratificat}
\alias{proporcio_grups_estratificat}
\title{proporcio_grups_estratificat}
\usage{
proporcio_grups_estratificat(
  dt = dades,
  factor.Y = "canvi612M.glicadaCAT2",
  grup = c("SEXE", "CODGLP1", "anys_DMcat4"),
  estratificat = "HBA1C_cat4"
)
}
\arguments{
\item{dt}{dt}

\item{factor.Y}{factor.Y}

\item{grup}{grup}

\item{estratificat}{estratificat}
}
\description{
proporcio_grups_estratificat
}
