% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{proporcions_grups}
\alias{proporcions_grups}
\title{Proporcions grups}
\usage{
proporcions_grups(
  dt = dades,
  factor = "canvi612M.glicadaCAT2",
  estrat = "SEXE"
)
}
\arguments{
\item{dt}{dt}

\item{factor}{factor}

\item{estrat}{estrat}
}
\description{
Proporcions grups
}
