% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{Pvalors_ajustats_compare}
\alias{Pvalors_ajustats_compare}
\title{Pvalors ajustats compare}
\usage{
Pvalors_ajustats_compare(
  objecte_compare = "T1.1.2",
  metodo = "BH",
  p = "p.overall",
  Sig = "No"
)
}
\arguments{
\item{objecte_compare}{objecte_compare}

\item{metodo}{metodo}

\item{p}{p}

\item{Sig}{Sig}
}
\description{
Pvalors ajustats compare
}
