% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{Pvalors_ajustats_taula}
\alias{Pvalors_ajustats_taula}
\title{Pvalors_ajustats_taula}
\usage{
Pvalors_ajustats_taula(
  objecte_taula = "OR.ajust",
  p.valors = "p valor",
  metodo = "BH"
)
}
\arguments{
\item{objecte_taula}{objecte_taula}

\item{p.valors}{p.valors}

\item{metodo}{metodo}
}
\description{
Pvalors_ajustats_taula
}
