% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{Pvalors_ajustats_Update_Compare}
\alias{Pvalors_ajustats_Update_Compare}
\title{Pvalors ajustats Update Compare}
\usage{
Pvalors_ajustats_Update_Compare(
  objecte_compare = "res",
  p = "p.overall",
  method = "BH"
)
}
\arguments{
\item{objecte_compare}{objecte_compare}

\item{p}{p}

\item{method}{method}
}
\description{
Pvalors ajustats Update Compare
}
