% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductorMacedonia.R
\name{read.conductor}
\alias{read.conductor}
\title{Read.conductor}
\usage{
read.conductor(fitxer, ...)
}
\arguments{
\item{fitxer}{Fitxer que el convertirem a tibble.}

\item{...}{Altres funcions}
}
\value{
Una taula tibble
}
\description{
Llegeix el fitxer conductor de diferents formats txt o rds o xls xlsx o data_frame tibble
}
\examples{
k1<-read.conductor(conductor1)
k1
}
