% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{reduccio_ajustada}
\alias{reduccio_ajustada}
\title{reduccio_ajustada}
\usage{
reduccio_ajustada(dt = dades, v.basal, v.final, mean.basal = NA)
}
\arguments{
\item{dt}{dt}

\item{v.basal}{v.basal}

\item{v.final}{v.final}

\item{mean.basal}{mean.basal}
}
\description{
reduccio_ajustada
}
