% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{regicor}
\alias{regicor}
\title{regicor}
\usage{
regicor(age, sex, smoker, diabetes, coltot, colhdl, sbp, dbp, divide = 1)
}
\arguments{
\item{age}{age}

\item{sex}{sex}

\item{smoker}{smoker}

\item{diabetes}{diabetes}

\item{coltot}{coltot}

\item{colhdl}{colhdl}

\item{sbp}{sbp}

\item{dbp}{dbp}

\item{divide}{divide}
}
\description{
regicor
}
