% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{resum3}
\alias{resum3}
\title{resum3}
\usage{
resum3(dt = dades, x = "val_last.HBA1C", estrat = "constant")
}
\arguments{
\item{dt}{dt}

\item{x}{x}

\item{estrat}{estrat}
}
\description{
resum3
}
