% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{resum_events_grup}
\alias{resum_events_grup}
\title{Resum events grups}
\usage{
resum_events_grup(
  d = "dadestotal",
  evento = "RD",
  temps = "TEMPS_RD2",
  grup = "sexe"
)
}
\arguments{
\item{d}{d}

\item{evento}{evento}

\item{temps}{temps}

\item{grup}{grup}
}
\description{
Resum events grups
}
