% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{resum_events_v2}
\alias{resum_events_v2}
\title{resum_events_v2}
\usage{
resum_events_v2(dades = dades, evento = "RD", temps = "temps")
}
\arguments{
\item{dades}{dades}

\item{evento}{evento}

\item{temps}{temps}
}
\description{
resum_events_v2
}
