% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{resum_events_v3}
\alias{resum_events_v3}
\title{Resum events v3}
\usage{
resum_events_v3(
  dt = "dadestotal",
  evento = "RD",
  temps = "temps",
  valorevent = "Si"
)
}
\arguments{
\item{dt}{dt}

\item{evento}{evento}

\item{temps}{temps}

\item{valorevent}{valorevent}
}
\description{
Resum events v3
}
