% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{resum_quanti_estrat}
\alias{resum_quanti_estrat}
\title{resum_quanti_estrat}
\usage{
resum_quanti_estrat(
  dt = dades,
  y = "valor_basal.GLICADA",
  grup = "CODGLP1",
  estrat = "HBA1C_cat4"
)
}
\arguments{
\item{dt}{dt}

\item{y}{y}

\item{grup}{grup}

\item{estrat}{estrat}
}
\description{
resum_quanti_estrat
}
