% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{Resum_taxa_incidencia_idp}
\alias{Resum_taxa_incidencia_idp}
\title{Resum taxa incidencia idp}
\usage{
Resum_taxa_incidencia_idp(
  dt = dades,
  evento = "event_tbc",
  temps = "anys_lliure_tbc",
  valorevent = "1",
  ...
)
}
\arguments{
\item{dt}{dt}

\item{evento}{evento}

\item{temps}{temps}

\item{valorevent}{valorevent}

\item{...}{Altres parametres}
}
\description{
Resum taxa incidencia idp
}
