% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{retorn_prediccio_MI}
\alias{retorn_prediccio_MI}
\title{retorn_prediccio_MI}
\usage{
retorn_prediccio_MI(
  data_imp = tempData,
  x = "HBpreADD",
  y = "canvi_ADD",
  dades_origen = dades
)
}
\arguments{
\item{data_imp}{data_imp}

\item{x}{x}

\item{y}{y}

\item{dades_origen}{dades_origen}
}
\description{
retorn_prediccio_MI
}
