% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{retorn_prediccio_MI_STR}
\alias{retorn_prediccio_MI_STR}
\title{retorn_prediccio_MI_STR}
\usage{
retorn_prediccio_MI_STR(
  data_imp = tempData,
  x = "HBpreADD",
  y = "canvi_ADD",
  dades_origen = dades,
  valor_subset = "<8",
  var_subset = "HBpreADD"
)
}
\arguments{
\item{data_imp}{data_imp}

\item{x}{x}

\item{y}{y}

\item{dades_origen}{dades_origen}

\item{valor_subset}{valor_subset}

\item{var_subset}{var_subset}
}
\description{
retorn_prediccio_MI_STR
}
