% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{retorn_prediccio_MI_STR2}
\alias{retorn_prediccio_MI_STR2}
\title{Retorn prediccio MI STR2}
\usage{
retorn_prediccio_MI_STR2(
  data_imp = tempData,
  x = "HBpreADD",
  y = "canvi_ADD",
  dades_origen = dades,
  valor_subset1 = ">=8",
  valor_subset2 = "<=10",
  var_subset = "HBpreADD"
)
}
\arguments{
\item{data_imp}{data_imp}

\item{x}{x}

\item{y}{y}

\item{dades_origen}{dades_origen}

\item{valor_subset1}{valor_subset1}

\item{valor_subset2}{valor_subset2}

\item{var_subset}{var_subset}
}
\description{
Retorn prediccio MI STR2
}
