% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDensityIncidence.R
\name{riskSetMatch}
\alias{riskSetMatch}
\title{Funcio que serveix per fer MATCHING.}
\usage{
riskSetMatch(
  ptid,
  event,
  terms,
  dat,
  Ncontrols,
  oldevent = "oldevent",
  caseid = "caseid",
  reuseCases = FALSE,
  reuseControls = FALSE,
  caseIndex = NULL,
  controlIndex = NULL,
  NoIndex = FALSE,
  cores = 2,
  dateterms = NULL
)
}
\arguments{
\item{ptid}{ptid}

\item{event}{event}

\item{terms}{terms}

\item{dat}{dat}

\item{Ncontrols}{Ncontrols}

\item{oldevent}{oldevent}

\item{caseid}{caseid}

\item{reuseCases}{reuseCases}

\item{reuseControls}{reuseControls}

\item{caseIndex}{caseIndex}

\item{controlIndex}{controlIndex}

\item{NoIndex}{NoIndex}

\item{cores}{cores}

\item{dateterms}{dateterms}
}
\value{
match
}
\description{
Aquestes funcions s adapten a les dades del registre danes. Dissenyat per ser utilitzat amb Statistics Denmark
}
\examples{
\dontrun{

#s ha d instalar aquest package, si no , no funciona :
#githubinstall::githubinstall("heaven",ref = "964bbbd",force = T)


case <- c(rep(0,40),rep(1,15))
ptid <- paste0("P",1:55)
sex <- c(rep("fem",20),rep("mal",20),rep("fem",8),rep("mal",7))
byear <- c(rep(c(2020,2030),20),rep(2020,7),rep(2030,8))
case.Index <- c(seq(1,40,1),seq(5,47,3))
control.Index <- case.Index
diabetes <- seq(2,110,2)
heartdis <- seq(110,2,-2)
diabetes <- c(rep(1,55))
heartdis <- c(rep(100,55))
library(data.table)

dat <- data.table::data.table(case,ptid,sex,byear,diabetes,heartdis,case.Index,control.Index)
# Very simple match without reuse - no dates to control for

out <- riskSetMatch("ptid",event="case",c("byear","sex"),dat,2,NoIndex=TRUE)

dat
out

# Risk set matching without reusing cases/controls -
# Some cases have no controls
out2 <- riskSetMatch("ptid",event="case",c("byear","sex"),dat,2,caseIndex="case.Index",
controlIndex="control.Index")

dat
out2

}
}
