% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDensityIncidence.R
\name{selecciona_parells}
\alias{selecciona_parells}
\title{selecciona parells}
\usage{
selecciona_parells(
  dt = "dt_poblacio",
  eventcontrol = T,
  reemplacement = F,
  Ncontrols = 1
)
}
\arguments{
\item{dt}{dt}

\item{eventcontrol}{eventcontrol}

\item{reemplacement}{reemplacement}

\item{Ncontrols}{Ncontrols}
}
\value{
parelles
}
\description{
selecciona parells
}
