% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{variables.ajust}
\alias{variables.ajust}
\title{Variables ajust}
\usage{
variables.ajust(x = "taula1", variables = variables)
}
\arguments{
\item{x}{x}

\item{variables}{variables}
}
\description{
Variables ajust
}
