% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{boxplot_variables_grup}
\alias{boxplot_variables_grup}
\title{boxplot_variables_grup}
\usage{
boxplot_variables_grup(
  dt = dades,
  variables = "OFT_WORST",
  grup = "DM",
  taulavariables = "variables_R.xls"
)
}
\arguments{
\item{dt}{dt}

\item{variables}{variables}

\item{grup}{grup}

\item{taulavariables}{taulavariables}
}
\description{
boxplot_variables_grup
}
