% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{extreure_OR}
\alias{extreure_OR}
\title{extreure_OR}
\usage{
extreure_OR(
  formu = "AnyPlaqueBasal~CD5L",
  dades = "dt",
  conditional = F,
  strata = "caseid"
)
}
\arguments{
\item{formu}{formu}

\item{dades}{dades}

\item{conditional}{conditional}

\item{strata}{strata}
}
\description{
extreure_OR
}
