% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{formula}
\alias{formula}
\title{FORMULA A PARTIR DE VARIABLES}
\usage{
formula(x = "taula1", y = "grup", eliminar = c("idp", "y"))
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{eliminar}{eliminar}
}
\value{
FORMULA
}
\description{
FORMULA A PARTIR DE VARIABLES
}
