% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDensityIncidence.R
\name{match_density_incidence}
\alias{match_density_incidence}
\title{Match density incidence}
\usage{
match_density_incidence(
  dt = "dt_poblacio",
  id = "idp",
  llistaPS = c("sexe"),
  eventcontrol = F,
  reemplacement = F,
  numcores = NA,
  Ncontrols = 1,
  seed = 123
)
}
\arguments{
\item{dt}{dt}

\item{id}{id}

\item{llistaPS}{llistaPS}

\item{eventcontrol}{eventcontrol}

\item{reemplacement}{reemplacement}

\item{numcores}{numcores}

\item{Ncontrols}{Ncontrols}

\item{seed}{seed}
}
\value{
match
}
\description{
Funcio que serveix per fer MATCHING.
}
\examples{
\dontrun{
dat

dat2<-match_density_incidence(dt=dat_test,
id="idp",llistaPS=c("sex"),
eventcontrol=FALSE,
reemplacement=FALSE,
numcores=NA,
Ncontrols=1,
seed=123)

dat2
}
}
