% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{plotKM_Incidence}
\alias{plotKM_Incidence}
\title{Plot KM_Incidence}
\usage{
plotKM_Incidence(
  y = "exitus.surv",
  grup = "grup",
  d = "dades",
  caption = "",
  llegenda = c("No", "Yes")
)
}
\arguments{
\item{y}{y}

\item{grup}{grup}

\item{d}{d}

\item{caption}{caption}

\item{llegenda}{legenda}
}
\description{
Plot KM_Incidence
}
