% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{resum_events}
\alias{resum_events}
\title{Resum events}
\usage{
resum_events(
  dades = "dadestotal",
  evento = "RD",
  temps = "temps",
  valorevent = "Si"
)
}
\arguments{
\item{dades}{dades}

\item{evento}{evento}

\item{temps}{temps}

\item{valorevent}{valorevent}
}
\description{
Resum events
}
