% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionsMacedonia.R
\name{resum_quanti}
\alias{resum_quanti}
\title{resum_quanti}
\usage{
resum_quanti(dt = dades, y = "valor_basal.GLICADA", grup = "constant")
}
\arguments{
\item{dt}{dt}

\item{y}{y}

\item{grup}{grup}
}
\description{
resum_quanti
}
