% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDensityIncidence.R
\name{riskSetMatch}
\alias{riskSetMatch}
\title{riskSetMatch}
\usage{
riskSetMatch(
  ptid,
  event,
  terms,
  dat,
  Ncontrols,
  oldevent = "oldevent",
  caseid = "caseid",
  reuseCases = FALSE,
  reuseControls = FALSE,
  caseIndex = NULL,
  controlIndex = NULL,
  NoIndex = FALSE,
  cores = 1,
  dateterms = NULL
)
}
\arguments{
\item{ptid}{ptid}

\item{event}{event}

\item{terms}{terms}

\item{dat}{dat}

\item{Ncontrols}{Ncontrols}

\item{oldevent}{oldevent}

\item{caseid}{caseid}

\item{reuseCases}{reuseCases}

\item{reuseControls}{reuseControls}

\item{caseIndex}{caseIndex}

\item{controlIndex}{controlIndex}

\item{NoIndex}{NoIndex}

\item{cores}{cores}

\item{dateterms}{dateterms}
}
\value{
match
}
\description{
riskSetMatch
}
